Subscription function complete and persisting

+++ profiles/views.py
@@ -14,35 +14,44 @@ def profile(request):
     """ Display the user's profile. """
     subscribed = False
     profile = get_object_or_404(UserProfile, user=request.user)
-    # Get value to subscriber toggle on profile page
-    subTog = request.POST.get('subTogBtn', '')
     # Get user instance and user email
     user = get_object_or_404(User, username=request.user)
     user_email = str(user.email)
-    if SubscriberList.objects.filter(email=user_email).exists():
-        subTog == "on"
-
-    if subTog == "on":
-        print("User wants to subscribe")
-        # if the user is already subscribed
+    if request.POST:
+        # POST Request so providing data
+        subTog = request.POST.get('subTogBtn', '')
+        # check if the user has already subscribed,
+        # and if so set subscriber toggle to ON
         if SubscriberList.objects.filter(email=user_email).exists():
-            print("Already Subscribed")
+            # Subscriber Exists
+            if subTog == "on":
+                # Toggle also on, Already Subscribed
+                subscribed = True
+            else:
+                # Already exists but toggle off
+                # Delete subscriber instance.
+                subscriber_instance = get_object_or_404(SubscriberList,
+                                                        email=user_email)
+                subscriber_instance.delete()
+                subscribed = False
         else:
-            # create and save subscriber instance.
-            SubscriberList.objects.create(email=user_email)
-        subscribed = True
+            # Subscriber does not already Exist
+            if request.POST.get('subTogBtn', '') == "on":
+                # Toggle on, so set up Subscriber,
+                # create and save subscriber instance.
+                SubscriberList.objects.create(email=user_email)
+                subscribed = True
+            else:
+                # toggle off, so do nothing
+                subscribed = False
     else:
-        print("User does not want to Subscribe")
+        # GET Request, so just looking to view profile
         if SubscriberList.objects.filter(email=user_email).exists():
-            print("Subscribed. Now delete")
-            # Delete subscriber instance.
-            subscriber_instance = get_object_or_404(SubscriberList,
-                                                    email=user_email)
-            subscriber_instance.delete()
-            print("Delete Subscription")
+            # Subscriber exists
+            subscribed = True
         else:
-            print("No Subscription")
-        subscribed = False
+            # subscriber does not already Exist
+            subscribed = False
 
     if request.method == 'POST':
         form = UserProfileForm(request.POST, instance=profile)
